% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{addColonyToTheColonies}
\alias{addColonyToTheColonies}
\title{Adds a colony to the group of colonies}
\usage{
\method{addColonyToTheColonies}(colony, colonies)
}
\arguments{
\item{colony}{AlphaSimRBee Colony object}

\item{colonies}{AlphaSimRBee Colonies object containing a list of colonies}
}
\value{
Updated AlphaSimRBee Colonies object

Updated AlphaSimRBee Colonies object
}
\description{
Adds a colony object into a grouped colonies object.
For example : A user can use this function to add a new colony to their apiary.
}
\examples{
# Create founder haplotypes
founderPop <- quickHaplo(nInd = 200, nChr = 1, segSites = 10)

# Set simulation parameters
SP <- SimParam$new(founderPop)

# Create population
pop <- newPop(founderPop, simParam = SP)

# Create colonies
apiary1 <- createMultipleMatedColonies(base, nColonies = 10, nAvgFathers = 15)
apiary1 <- buildUpColonies(apiary1, nWorkers = 1000, nDrones = 100)

# Creates colony
colony1 <- createColony(queen = base[1], fathers = base[2:15])
colony1 <- addWorkers(colony1, nInd = 2000)

# Add colony1 to apiary1
apiary1 <- addColonyToTheColonies(colony1, apiary1)
}
