% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{addVirginQueens}
\alias{addVirginQueens}
\title{Create additional virgin queens}
\usage{
\method{createVirginQueens}(colony, nVirginQueens)
}
\arguments{
\item{colony}{AlphaSimRBee Colony object}

\item{nVirginQueens}{Numeric. Number of virgin queens to create}
}
\value{
Updated AlphaSimRBee Colony object
}
\description{
Creates the specified number of virgin queens in the colony
      \by crossing the current queen and the fathers and adds them in
      \ the \code{colony@virgin_queens} slot.
}
\examples{
#Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
pop = newPop(founderPop, simParam=SP)

#Creates colony with queen and fathers
colony1 = createColony(queen = base[1], fathers = base[2:15])

#Insert workers and drones into the colony
colony1 = addWorkers(colony1, nInd = 2000)
colony1 = addDrones(colony, nInd = 100)

#Insert 10  virgin queens into the colony
colony1 = addVirginQueens(colony1, nInd = 10)

}
