% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{addWorkers}
\alias{addWorkers}
\title{Add workers to the colony}
\usage{
\method{addWorkers}(colony, nInd)
}
\arguments{
\item{colony}{AlphaSimRBee Colony object from the \code{createColony(...)} call}

\item{nInd}{Integer, number of workers to add.}
}
\value{
Updated AlphaSimRBee Colony object
}
\description{
Create workers and store them in the \code{colony@workers} slot. If there is
already some workers present in the hive, the function will not overwrite them but instead
combine the newly created and the existing workers. The function returns the updated colony.
}
\examples{
Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
pop = newPop(founderPop, simParam=SP)

#Creates colony
colony1 = createColony(queen = base[1], fathers = base[2:15])
colony1 = addWorkers(colony1, nInd = 2000)

}
