% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{buildUpColonies}
\alias{buildUpColonies}
\title{Build up Colonies by adding workers and drones}
\usage{
\method{buildUpColonies}(colonies, nWorkers, nDrones)
}
\arguments{
\item{colonies}{AlphaSimRBee Colonies object containing a list of colonies}

\item{nWorkers}{Desired number of workers wanted in the colonies}

\item{nDrones}{Desired number of drones wanted in the colonies}
}
\value{
An updated AlphaSimRBee Colonies object
}
\description{
Workers and drones are added to the colonies to build them
up to the number of desired workers and drones (nWorkers and nDrones).
For example: a user may build up colonies in the Period 1 and if events such as split or swarming
occur.
}
\examples{
#Create founder haplotypes
founderPop <- quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP <- SimParam$new(founderPop)

#Create population
base <- newPop(founderPop, simParam=SP)

#Create 10 virgin queen colonies
 apiary1 <- createMultipleMatedColonies(founderPop = base, nColonies = 10, nAvgFathers = 15)

#Build up colonies by adding 1000 workers and 100 drones to each colony in the "colonies" list
 apiary1 <- buildUpColonies(apiary1, nWorkers = 1000, nDrones = 100)

}
