% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{collapseColonies}
\alias{collapseColonies}
\title{collapseColonies}
\usage{
\method{collapseColonies}(colonies, ID)
}
\arguments{
\item{colonies}{AlphaSimRBee Colonies object containing a list of colonies}

\item{ID}{IDs of "colony" class objects listed in the "colonies" object}
}
\value{
An updated AlphaSimRBee Colonies object
}
\description{
:
}
\examples{

}
