% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{Colonies-class}
\alias{Colonies-class}
\title{Colonies}
\arguments{
\item{x}{Colonies, colonies}

\item{i}{numeric or character, index or name to select a colony}
}
\description{
The Colonies represents a set of colonies. It behaves like a list object.
}
\section{Slots}{

\describe{
\item{\code{colonies,}}{a collection of \code{\link{Colony-class}} objects}
}}

\seealso{
\code{\link{selectColonies}}, \code{\link{addColonyToColonies}}
}
