% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{computeQueenAge}
\alias{computeQueenAge}
\title{Compute the queen's age in years}
\usage{
\method{computeQueenAge}(colony, year)
}
\arguments{
\item{currentYear}{Integer, current year}

\item{colony}{AlphaSimRBee Colony object from the \code{createColony(...)} call}
}
\value{
Integer, the age of the queen.
}
\description{
Compute the age of the queen from the \code{colony@queen@misc$yearOfBirth} slot
}
\examples{
#Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
pop = newPop(founderPop, simParam=SP)

#Creates colony
colony1 = createColony(queen = base[1], fathers = base[2:15])
setQueenAge(colony, year = 1)
extractQueenYOB(colony)
computerQueenAge(colony, currentYear = 5)

}
