% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{createFounderDrones}
\alias{createFounderDrones}
\title{Creates drones from base population}
\usage{
\method{createFounderDrones}(queenPop, nDronesPerQueen)
}
\arguments{
\item{queenPop}{AlphaSimRBee Colony object from the \code{createColony(...)}}

\item{nDronesPerQueen}{Integer, number of drones to create}
}
\value{
AlphaSim population object of created drones.
}
\description{
Creates population of drones from base population.
      \Drones are created as double haploids. Founder drones are used when crating colonies to be used as fathers.
}
\examples{
#Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
pop = newPop(founderPop, simParam=SP)

founderDrones = createFounderDrones(pop[2:110], nDronesPerQueen = 100)
colony1 = createColony(queen = pop[1], fathers = founderDrones[1:17])

}
