% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{createColonies}
\alias{createColonies}
\title{}
\usage{
\method {createColonies}(..., n)
}
\arguments{
\item{...}{one or more \code{\link{Colony-class}} and/or
\code{\link{Colonies-class}} objects.}

\item{n}{Number of colony-class objects to add to the colonies-class object}
}
\value{
Returns an empty object of \code{\link{Colonies-class}}
}
\description{
Creates a new \code{\link{Colonies-class}} from one or more
\code{\link{Colony-class}} and/or \code{\link{Colonies-class}}
objects.
}
