% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{createDCA}
\alias{createDCA}
\title{Create a drone congregation area (DCA)}
\usage{
createDCA(x, nInd = NULL)
}
\arguments{
\item{x}{Colony or Colonies}

\item{nInd}{numeric, number of drones to access, if NULL all drones
are accessed, otherwise a random sample}
}
\value{
Pop, population object with drones
}
\description{
Create a drone congregation area (DCA) from colony or colonies.
}
\examples{
# AlphaSimR
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 10)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)

# Honeybee
drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addDrones(colony1, nInd = 10)
colony2 <- addDrones(colony2, nInd = 20)
createDCA(colony1)
createDCA(colony2)
createDCA(colony2, nInd = 10)@id
createDCA(colony2, nInd = 10)@id

apiary <- c(colony1, colony2)
createDCA(apiary)
createDCA(apiary, nInd = 10)

}
