% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{createDrones}
\alias{createDrones}
\title{Creates drones of the colony as double haploids}
\usage{
\method{createDrones}(colony, nInd)
}
\arguments{
\item{colony}{AlphaSimRBee Colony object from the \code{createColony(...)} call}

\item{nInd}{Integer, the umber of drones to create.}
}
\value{
AlphaSim population object of created drones.
}
\description{
Creates the specified number of drones in the colony
      \as double haploids from the current queen \code{colony@queen}.
}
\examples{
# Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
pop = newPop(founderPop, simParam=SP)

#Creates colony
founderDrones = createFounderDrones(pop[3:200], nDronesPerQueen = 17)
colony1 = createColony(queen = pop[1], fathers = founderDrones[1:17])

#Create drones
colony1@drones = createDrones(colony1, nInd = 200)

}
