% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{createMultipleMatedColonies}
\alias{createMultipleMatedColonies}
\title{Create a list object of class "colonies" containing mated queens, virgin queens and fathers}
\usage{
\method{createMultipleMatedColonies}(founderPop, nColonies, nAvgFathers)
}
\arguments{
\item{founderPop}{The initial founder population}

\item{nColonies}{Number of colonies the use wants to create}

\item{nAvgFathers}{Average number of fathers that mates with a queen}
}
\value{
A AlphaSimRBee Colonies object
}
\description{
The function is intended for creating initial colonies from
'FOUNDERPOP'. The user can create a list containing their desired number of colonies (nColonies).
The colonies created contain a mated queen, a virgin queen and fathers (varying number surrounding the nAvgFathers)
}
\examples{
#Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
base = newPop(founderPop, simParam=SP)

#Create 10 virgin queen colonies
 apiary1 = createMultipleMatedColonies(founderPop = base, nColonies = 10, nAvgFathers = 15)

}
