% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{createVirginQueens}
\alias{createVirginQueens}
\title{Creates virgin queen}
\usage{
\method{createVirginQueens}(colony, nInd)
}
\arguments{
\item{colony}{AlphaSimRBee Colony object from the \code{createColony(...)} call}

\item{nInd}{Integer, the number of virgin queens to create.}
}
\value{
AlphaSim population object of created virgin queens.
}
\description{
Creates the specified number of virgin queens in the colony
       \by mating the current queen and the fathers in the \code{colony@queen@misc$fathers} slot.
}
\examples{
# Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
pop = newPop(founderPop, simParam=SP)

#Creates colony
founderDrones = createFounderDrones(pop[3:200], nDronesPerQueen = 17)
colony1 = createColony(queen = pop[1], fathers = founderDrones[1:17])

#Crate virgin queens
colony1@virgin_queens = createVirginQueens(colony1, nInd = 17)

}
