% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{createWorkers}
\alias{createWorkers}
\title{Creates workers of the colony}
\usage{
\method{createWorkers}(colony, nInd)
}
\arguments{
\item{colony}{AlphaSimRBee Colony object from the \code{createColony(...)} call}

\item{nInd}{Integer, number of workers to create}
}
\value{
AlphaSim population object of created workers.
}
\description{
Creates the specified number of workers in the colony
      \by mating the current queen and the fathers in the \code{colony@queen@misc$fathers} slot.
}
\examples{
#Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
pop = newPop(founderPop, simParam=SP)

#Creates colony
founderDrones = createFounderDrones(pop[3:200], nDronesPerQueen = 17)
colony1 = createColony(queen = pop[1], fathers = founderDrones[1:17])

#Creat workers
colony1@workers = createWorkers(colony1, nInd = 1000)

}
