% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{crossColonies}
\alias{crossColonies}
\title{crossColonies}
\usage{
\method{crossColonies}(colonies, DCA, nAvgFathers, crossVirginQueen, fathers, pWorkers, pDrones)
}
\arguments{
\item{colonies}{AlphaSimRBee Colonies object containing a list of colonies}

\item{nAvgFathers}{}
}
\value{
An updated AlphaSimRBee Colonies object
}
\description{
:  Crosses colonies with a virgin queen to a group of fathers pulled from the DCA
\creates workers, drones and a new virgin queen and write them to the corresponding
\slots of the colonies object.
#IF the colonies are queen-less - select a queen from the virgin queen - if not, mate the current queen!!!
}
\examples{
#Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
base = newPop(founderPop, simParam=SP)

#Create 10 virgin queen colonies
 apiary1 = createMultipleMatedColonies(founderPop = base, nColonies = 10, nAvgFathers = 15)

#Build up colonies by adding 1000 workers and 100 drones to each colony in the "colonies" list
 apiary1 = buildUpColonies(apiary1, nWorkers = 1000, nDrones = 100)

 TODO FINISH

}
