% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{crossVirginQueen}
\alias{crossVirginQueen}
\title{Crosses a virgin queen to a group drones.}
\usage{
\method{crossVirginQueen}(virginQueen, fathers)
}
\arguments{
\item{virginQueen}{AlphaSimR population object}

\item{fathers}{AlphaSimR population class.}
}
\value{
AlphaSim population object of a mated queen
}
\description{
Crosses a virgin queen to a group of drones
}
\examples{
# Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
pop = newPop(founderPop, simParam=SP)

#Creates colony
founderDrones = createFounderDrones(pop[3:200], nDronesPerQueen = 17)
colony1 = createColony(queen = pop[1], fathers = founderDrones[1:17])
colony2 = createColony(queen = pop[2], fathers = founderDrones[18:37])

#Create drones that will be in DCA
colony1@drones = createDrones(colony1, nInd = 300)
colony2@drones = createDrones(colony2, nInd = 300)

DCA = createDCA(c(colony1, colony2))

#pull drone packages from DCA
dronePack = pullDronePackagesFromDCA(DCA, n = 7, nAvgFathers = 19)

# Cross the virgin queen

virginQueen = colony1@virgin_queens
queen = crossVirginQueen(virginQueen, fathers = dronePack[4])

}
