% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{extractQueenYOB}
\alias{extractQueenYOB}
\title{Extract the queen's year of birth}
\usage{
\method{extractQueenYOB}(colony)
}
\arguments{
\item{colony}{AlphaSimRBee population object of class "Colony"}
}
\value{
Integer, the year of birth of the queen.
}
\description{
Extract the queen's year of birth \code{colony@queen@misc$yearOfBirth} slot
}
\examples{
#Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
pop = newPop(founderPop, simParam=SP)

#Creates colony
colony1 = createColony(queen = base[1], fathers = base[2:15])
setQueenAge(colony, 1)
extractQueenYOB(colony)

}
