% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getCsdGeno}
\alias{getCsdGeno}
\title{Get the Genotypes of the csd locus}
\usage{
\method{getCsdGeno}(x, csd)
}
\arguments{
\item{x}{Undefined argument. Can be a "Colony" class or "Colonies" class}

\item{csd}{complementary sex determination locus- list with nodes chr, start and stop}
}
\value{

}
\description{

}
\examples{

}
