% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{getFathers}
\alias{getFathers}
\title{Access fathers}
\usage{
getFathers(x, nInd = NULL)
}
\arguments{
\item{x}{Colony or Colonies}

\item{nInd}{numeric, number of fathers to access, if NULL all fathers
are accessed, otherwise a random sample}
}
\value{
When \code{x} is Colony then return is Pop, population object with fathers
When \code{x} is Colonies then return is a list of Pop, population objects with fathers
}
\description{
Access fathers (drones the queen mated with)
}
\examples{
# AlphaSimR
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 10)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)

# Honeybee
drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
getFathers(colony1)
getFathers(colony1)@id
drones@id

apiary <- c(colony1, colony2)
getFathers(apiary)
getFathers(apiary)[[1]]@id
getFathers(apiary)[[2]]@id

}
