% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{getIDs}
\alias{getIDs}
\title{Get the colonies IDs from the colonies}
\usage{
\method{getIDs}(colonies)
}
\arguments{
\item{colonies}{AlphaSimRBee Colonies object from the \code{createColonies(...)} call}
}
\value{
Character. Colony IDs
}
\description{
Get the colony IDs from the colonies
}
\examples{
 #Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
pop = newPop(founderPop, simParam=SP)

#Create an apiary containing 10 colonies
Apiary1 = createColonies(,n = 10)

#Get colony IDs from the colonies
getIDs (Apiary1)

}
