% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{getQueen}
\alias{getQueen}
\title{Access the queen}
\usage{
getQueen(x)
}
\arguments{
\item{x}{Colony or Colonies}
}
\value{
When \code{x} is Colony then return is Pop, population object with the queen
When \code{x} is Colonies then return is a list of Pop, population objects with the queen
}
\description{
Access the queen of a colony
}
\examples{
# AlphaSimR
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 10)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)

# Honeybee
drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
getQueen(colony1)
getQueen(colony2)

apiary <- c(colony1, colony2)
getQueen(apiary)

}
