% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{getVirginQueens}
\alias{getVirginQueens}
\title{Access virgin queens}
\usage{
getVirginQueens(x, nInd = NULL)
}
\arguments{
\item{x}{Colony or Colonies}

\item{nInd}{numeric, number of virgin queens to access, if NULL all virgin queens
are accessed, otherwise a random sample}
}
\value{
When \code{x} is Colony then return is Pop, population object with virgin queens
When \code{x} is Colonies then return is a list of Pop, population objects with virgin queens
}
\description{
Access virgin queens
}
\examples{
# AlphaSimR
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 10)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)

# Honeybee
drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addVirginQueens(colony1, nVirginQueens = 1)
colony2 <- addVirginQueens(colony1, nVirginQueens = 10)
getVirginQueens(colony1)
getVirginQueens(colony2)
getVirginQueens(colony2, nInd = 2)
getVirginQueens(colony2, nInd = 2)@id
getVirginQueens(colony2, nInd = 2)@id

apiary <- c(colony1, colony2)
getVirginQueens(apiary)
getVirginQueens(apiary)[[1]]@id
getVirginQueens(apiary)[[2]]@id

}
