% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nCaste}
\alias{nCaste}
\title{Number of individuals of a caste in a colony}
\usage{
nCaste(x, caste)
}
\arguments{
\item{x}{Colony or Colonies}

\item{caste}{character, "queen", "fathers", "virgin_queens", "workers", or "drones"}
}
\value{
Integer, named by colony id when \code{x} is Colonies
}
\description{
Returns the number of individuals of a caste in a colony
}
\examples{
AlphaSimR
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 10)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)

# Honeybees
drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addWorkers(colony1, nInd = 100)
colony1 <- addDrones(colony1, nInd = 10)
colony1 <- addVirginQueens(colony1, nInd = 3)
colony2 <- addWorkers(colony2, nInd = 200)
nCaste(colony1, caste = "queen")
nCaste(colony1, caste = "fathers")
nCaste(colony1, caste = "virgin_queens")
nCaste(colony1, caste = "workers")
nCaste(colony1, caste = "drones")

apiary <- c(colony1, colony2)
nCaste(apiary, caste = "queen")
nCaste(apiary, caste = "fathers")
nCaste(apiary, caste = "virgin_queens")
nCaste(apiary, caste = "workers")
nCaste(apiary, caste = "drones")

}
