% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nColonies}
\alias{nColonies}
\title{Returns number of colonies in the colonies object}
\usage{
\method{nColonies}(colonies)
}
\arguments{
\item{colonies}{AlphaSimRBee Colonies object from the \code{createColonies(...)} call}
}
\value{
Integer. Number of colonies present in the colonies AlphaSimRBee object
}
\description{
Returns the number of colonies present in the colonies object
}
\examples{
#Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
pop = newPop(founderPop, simParam=SP)

# Create an apiary containing 10 colonies
Apiary1 = createColonies(,n = 10)

#Check number of colonies present in the apiary
nColonies(apiary2)

}
