% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nDrones}
\alias{nDrones}
\title{Number of drones in a colony}
\usage{
nDrones(x)
}
\arguments{
\item{x}{Colony or Colonies}
}
\value{
Integer, named by colony id when \code{x} is Colonies
}
\description{
Returns the number of drones in a colony
}
\examples{
AlphaSimR
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 10)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)

# Honeybees
drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addDrones(colony1, nInd = 100)
colony2 <- addDrones(colony2, nInd = 200)
nDrones(colony1)
nDrones(colony2)

apiary <- c(colony1, colony2)
nDrones(apiary)

}
