% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nFathers}
\alias{nFathers}
\title{Number of fathers queens present in the colony}
\usage{
\method{nFathers}(colony)
}
\arguments{
\item{colony}{AlphaSimRBee population object of class "Colony"}
}
\value{
Integer. Number of fathers present in the colony AlphaSimRBee object
}
\description{
Returns the number of fathers present in the colony object.
If no queens are present in the colony, no fathers are present and function is returned as 0.
}
\examples{
#' #Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
pop = newPop(founderPop, simParam=SP)

#Creates colony
colony1 = createColony(queen = base[1], fathers = base[2:15])

Check number of fathers in the colony
nFathers(colony1)

}
