% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nQueen}
\alias{nQueen}
\title{Number of queens in a colony}
\usage{
nQueen(x)
}
\arguments{
\item{x}{Colony or Colonies}
}
\value{
Integer, named by colony id when \code{x} is Colonies
}
\description{
Returns the number of queens in a colony (expect 0 or 1)
}
\examples{
AlphaSimR
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 10)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)

# Honeybees
drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:8])
nQueen(colony1)
nQueen(colony2)
colony2 <- removeQueen(colony2)
nQueen(colony2)

apiary <- c(colony1, colony2)
nQueen(apiary)

}
