% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nVirginQueens}
\alias{nVirginQueens}
\title{Number of virgin queens present in the colony}
\usage{
\method{nVirginQueens}(colony)
}
\arguments{
\item{colony}{AlphaSimRBee population object of class "Colony"}
}
\value{
Integer. Number of virgin queens present in the colony AlphaSimRBee object
}
\description{
Returns the number of virgin queens present in the colony object
}
\examples{
#' #Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
pop = newPop(founderPop, simParam=SP)

#Creates colony
colony1 = createColony(queen = base[1], fathers = base[2:15])
colony1@virgin_queens = createVirginQueens(colony1, nInd = 5)

Check number of drones in the colony
nVirginQueens(colony1)

}
