% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nWorkers}
\alias{nWorkers}
\title{Number of workers present in the colony}
\usage{
\method{nWorkers}(colony)
}
\arguments{
\item{colony}{AlphaSimRBee population object of class "Colony"}
}
\value{
Integer. Number of workers present in the colony AlphaSimRBee object
}
\description{
Returns the number of workers present in the colony object
}
\examples{
#' #Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
pop = newPop(founderPop, simParam=SP)

#Creates colony
colony1 = createColony(queen = base[1], fathers = base[2:15])
colony1@workers = createWorkers(colony1, nInd = 1000)

Check number of workers in the colony
nWorkers(colony1)

}
