% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{pullColonies}
\alias{pullColonies}
\title{Pull the colonies from the colony list based on IDs.}
\usage{
\method{pullColonies}(colonies,ID, p)
}
\arguments{
\item{colonies}{AlphaSimRBee Colonies object containing a list of colonies}

\item{ID}{IDs of "colony" class objects listed in the "colonies" object}

\item{p}{Percentage of colonies to pull out of the "colonies" list object}
}
\value{
Two lists: a list of selected colonies and an updated input colonies
}
\description{
Pull the colonies from the list of all colonies based
on colony IDs and return two lists: a list of selected colonies and
updated original colonies
}
\examples{
 # Create founder haplotypes

founderPop <- quickHaplo(nInd=300, nChr=1, segSites=10)

# Set simulation parameters

SP <- SimParam$new(founderPop)

# Create population

pop <- newPop(founderPop, simParam=SP)

# Create colonies

founderDrones <- createFounderDrones(pop[13:300], nDronesPerQueen = 17)
colony1 <- createColony(queen = pop[1], fathers = founderDrones[1:17])
colony2 <- createColony(queen = pop[2], fathers = founderDrones[18:37])
colony3 <- createColony(queen = pop[3], fathers = founderDrones[37:51])

# Put the colonies together to the apiary

apiary <- c(colony1, colony2, colony3)

# Pull the selected colonies

pulledColonies <- pullColonies(colonies = apiary, ID = c(1,2))

apiary <- pulledColonies$remainingColonies
pull <- pulledColonies$pulledColonies

}
