% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{pullDronePackagesFromDCA}
\alias{pullDronePackagesFromDCA}
\title{Pulls a drone package from DCA}
\usage{
\method{pullDronePackagesFromDCA}(DCA, n, nAvgFathers)
}
\arguments{
\item{DCA}{object of class pop. AlphaSimRBee Colony object from the \code{createColony(...)} call}

\item{n}{Integer. Number of the packages that needs to be created.}

\item{nAvgFathers}{Integer. The average number of drones that will be in the package.}
}
\value{
Two AlphaSim population objects of the colony and the group of pulled individuals.
}
\description{
Pulls the packages of drones from DCA. Each package is then used to cros colonies.
             Number of drones in package is sampled from the Poisson distribution
             with the average = nAvgFathers. Selected drones are removed from DCA.
}
\examples{
# Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
pop = newPop(founderPop, simParam=SP)

#Creates colony
founderDrones = createFounderDrones(pop[3:200], nDronesPerQueen = 17)
colony1 = createColony(queen = pop[1], fathers = founderDrones[1:17])
colony2 = createColony(queen = pop[2], fathers = founderDrones[18:37])

#Create drones that will be in DCA
colony1@drones = createDrones(colony1, nInd = 300)
colony2@drones = createDrones(colony2, nInd = 300)

DCA = createDCA(c(colony1, colony2))

#pull drone packages from DCA
dronePack = pullDronePackagesFromDCA(DCA, n = 7, nAvgFathers = 19)

}
\seealso{
\code{\link[??????]{pullIndFromCaste}}
}
