% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{pullDronesFromDCA}
\alias{pullDronesFromDCA}
\title{Pulls the drones from the DCA}
\usage{
\method{pullDronesFromDCA}(DCA, nInd)
}
\arguments{
\item{DCA}{AlphaSimR population object created with \code{createDCA(...)} call}

\item{nInd}{Integer, the number of drones to pull from the DCA}
}
\value{
A list with two elements. The first element is the AlphaSimR population object of
selected drones. The second element is the updated DCA with the selected drones removed.
}
\description{
Pulls a specified number of drones from the DCA and updates the DCA.
              Selected drones are removed from DCA.
}
\examples{
# Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
pop = newPop(founderPop, simParam=SP)

#Creates colony
founderDrones = createFounderDrones(pop[3:200], nDronesPerQueen = 17)
colony1 = createColony(queen = pop[1], fathers = founderDrones[1:17])
colony2 = createColony(queen = pop[2], fathers = founderDrones[18:37])

#Create drones that will be in DCA
colony1@drones = createDrones(colony1, nInd = 300)
colony2@drones = createDrones(colony2, nInd = 300)

DCA = createDCA(c(colony1, colony2))

# Pull drones from DCA
fathers =  pullDronesFromDCA(DCA, nInd = 19)

}
