% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{pullIndFromCaste}
\alias{pullIndFromCaste}
\title{Randomly pulls a number of individuals from any caste group.}
\usage{
\method{pullIndFromCaste}(colony, caste, nInd)
}
\arguments{
\item{colony}{Colony class. AlphaSimRBee Colony object from the \code{createColony(...)} call}

\item{caste}{Character. Replicating the caste class structure present in the hive (queen, drones, workers etc)}

\item{nInd}{Integer. Number of individuals to be pulled from the caste}
}
\value{
Two AlphaSim population objects of the colony and the group of pulled individuals.
}
\description{
Pulls and separates a random number of individuals from any caste group.
Two list groups are created, the group of pulled individuals and the colony.
}
\examples{
inst/examples/examples_pullIndFromCaste.R
Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
pop = newPop(founderPop, simParam=SP)

#Creates colony
founderDrones = createFounderDrones(pop[3:200], nDronesPerQueen = 17)
colony1 = createColony(queen = pop[1], fathers = founderDrones[1:17])

#Create drones
colony1@drones = createDrones(colony1, nInd = 300)

# Pull individuals from the caste
indDrone = pullIndFromCaste(colony1, caste = 'drones', nInd = 1)

}
\seealso{
\code{\link[??????]{pullIndFromCaste}}
}
