% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{pullIndFromPop}
\alias{pullIndFromPop}
\title{Pull individuals from a population}
\usage{
pullIndFromPop(pop, nInd)
}
\arguments{
\item{pop}{Pop}

\item{nInd}{numeric, number of individuals to pull}
}
\value{
Pop, population object with drones
}
\description{
Pull individuals from a population and update the population
}
\examples{
# AlphaSimR
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 10)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)
pullIndFromPop(basePop, nInd = 2)
pullIndFromPop(basePop, nInd = 3)

}
