% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{pullVirginQueens}
\alias{pullVirginQueens}
\title{Pull virgin queens from a colony}
\usage{
pullVirginQueens(x, nInd = NULL)
}
\arguments{
\item{x}{Colony or Colonies}

\item{nInd}{numeric, number of virgin queens to pull, if NULL all virgin queens
are pulled, otherwise a random sample}
}
\value{
When \code{x} is Colony then return is a list with a population object (Pop) with the pulled virgin queen and remaining colony
When \code{x} is Colonies then return is a list with a list of population object (Pop) with the pulled virgin queens and remaining colonies
}
\description{
Pull virgin queen from a colony
}
\examples{
# AlphaSimR
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 10)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)

# Honeybee
drones <- createFounderDrones(pop = basePop[1], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[2], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[3], fathers = drones[6:10])
colony1 <- addVirginQueens(colony1, nInd = 10)
pullVirginQueens(colony1)
pullVirginQueens(colony1, nInd = 2)
pullVirginQueens(colony2)

apiary <- c(colony1, colony2)
pullVirginQueens(apiary)
nVirginQueens(apiary)
nVirginQueens(pullVirginQueens(apiary)$colonies)

}
