% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{removeDrones}
\alias{removeDrones}
\title{Remove selected percentage of drones}
\usage{
\method{removeWorkers}(colony, nWorkers)
}
\arguments{
\item{colony}{Colony class. AlphaSimR Colony object from the \code{createColony(...)} call}

\item{p}{Numeric. 0<=p>=1 .}
}
\description{
To decrease the number of drones (for example, in winter)
}
\examples{
inst/examples/examples_removeDrones.R
#' @return Updated AlphaSimRBee Colony object
}
