% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{replaceDrones}
\alias{replaceDrones}
\title{Replaces drone with new drone with new genetic information}
\usage{
\method{replaceWorkers}(colony, p)
}
\arguments{
\item{colony}{AlphaSimRBee Colony object.}

\item{p}{Numeric, proportion of drones to be replaced with new ones.}
}
\value{
Updated AlphaSimRBee Colony object
}
\description{
Replace a proportion of drones in the new with new drones from the same queen.
A user would want to replace a proportion (or all) of drones after swarming and supersedure or
due to the short-life span of the drones.
}
\examples{
#Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
pop = newPop(founderPop, simParam=SP)

#Creates colony
colony1 = createColony(queen = base[1], fathers = base[2:15])
colony1 = addDrones(colony1, nInd = 2000)
colony1 = replaceDrones(colony1, p = 0.2)

}
