% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{replaceWorkersColonies}
\alias{replaceWorkersColonies}
\title{Replace workers in all the colonies of a Colonies object}
\usage{
replaceWorkersColonies(colonies)
}
\arguments{
\item{colonies}{AlphaSimRBee Colonies object containing a list of colonies}
}
\value{
An updated AlphaSimRBee Colonies object
}
\description{
The function replaces a given proportion of workers in all the colonies
of a Colonies object with new workers from the same mated queen. The default percentage
is set to 1, hence replacing all workers If some colonies do not have workers, the
function does not add them.
}
\examples{
#Create founder haplotypes
founderPop <- quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP <- SimParam$new(founderPop)

#Create population
base <- newPop(founderPop, simParam=SP)

#Create 10 virgin queen colonies
 apiary1 <- createMultipleMatedColonies(founderPop = base, nColonies = 10, nAvgFathers = 15)

#Build up colonies by adding 1000 workers and 100 drones to each colony in the "colonies" list
 apiary1 <- buildUpColonies(apiary1, nWorkers = 1000, nDrones = 100)

#Replace the workers
  apiary1 <- replaceWorkers(apiary1)

}
