% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{reQueenColonies}
\alias{reQueenColonies}
\title{reQueen Colonies}
\usage{
\method{reQueenColonies}(colonies, queens)
}
\arguments{
\item{colonies}{AlphaSimRBee Colonies object containing a list of colonies}

\item{queens}{Selected individuals to insert into the queen slot of the colony object}
}
\value{
An updated AlphaSimRBee Colonies object
}
\description{
: Add a new queen/virgin queen into the queens slot of the colonies.
For example: this can be used to re-queen swarmed or split colonies where no queens are present.
}
\examples{
#Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
base = newPop(founderPop, simParam=SP)

#Create 10 virgin queen colonies
 apiary1 = createMultipleMatedColonies(founderPop = base, nColonies = 10, nAvgFathers = 15)

#Build up colonies by adding 1000 workers and 100 drones to each colony in the "colonies" list
 apiary1 = buildUpColonies(apiary1, nWorkers = 1000, nDrones = 100)

 #Split all the colonies
 tmp = splitColonies(apiary1)
 apiary1 = tmp$remnants
 splitcolonies = tmp$splits

 #Create 10 virgin queens
 virginQueens = createVirginQueens(apiary1[[10]], nColonies(splitcolonies))

 # Requeen the splits
 splitcolonies = reQueenColonies(splitcolonies, queens = virginQueens)

}
