% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{resetEvents}
\alias{resetEvents}
\title{Reset the swarm, split, supersedure events}
\usage{
\method{resetEvents}(colony)
}
\arguments{
\item{colony}{AlphaSimRBee Colony object.}
}
\value{
An updated AlphaSimRBee Colony object
}
\description{
Reset the slots swarm, split and supersedure to FALSE.
A user will use this function at the end of a yearly cycle to reset the events,
allowing the user to track the events of the current year without overlap.
}
\examples{

}
