% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{selectColonies}
\alias{selectColonies}
\title{Select the colonies from the colony list based on IDs.}
\usage{
\method{selectColonies}(colonies, IDs)
}
\arguments{
\item{colonies}{AlphaSimRBee Colonies object containing a list of colonies}

\item{ID}{IDs of "colony" class objects listed in the "colonies" object}
}
\value{
A list of selected colonies.
}
\description{
Select the colonies from the list of all colonies based
on colony IDs and return a list of selected colonies.
}
\examples{
inst/examples/examples_selectColonies.R
#Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
base = newPop(founderPop, simParam=SP)

#Create an apiary with 10 mated colonies (contains only queens, fathers and virgin queens)
apiary1 = createMultipleMatedColonies(base, nColonies = 10, nAvgFathers = 15)

#Build up the colonies to full size by adding workers and drones (drones are 1/10 size of workers)
colonyFullSize = 1000
apiary1 = buildUpColonies(apiary1, nWorkers = colonyFullSize, nDrones = colonyFullSize * 0.1)

#Check colony IDs in colony
apiary@colonies

#Select colonies with IDs 8, 40 and 45
chosenColonies = selectColonies(apiary1, ID = c(8,40,45))

}
