% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{splitColony}
\alias{splitColony}
\title{Split the colony in two colonies.}
\usage{
\method{splitColony}(colony, pSplit, newQueen, crossVirginQueen, fathers, nWorkers, nDrones, splitLocation)
}
\arguments{
\item{colony}{Colony class. AlphaSimRBee Colony object from the \code{createColony(...)} call}

\item{pSplit}{Integer. Percentage of hive to split}

\item{newQueen}{AlphaSimR population object. A new mated queen is brought into the colony from other source}

\item{crossVirginQueen}{Logical. If no mated queen is introduced, a virgin queen must be present to mate and continue colony}

\item{fathers}{AlphaSimR population object. Number of fathers pulled from the DCA}

\item{pWorkers}{Numeric, proportion of workers that are replaced with the workers from the new queen in the split}

\item{splitLocation}{Integer. X,Y coordinates of newly made split hive}
}
\value{
Two AlphaSim population objects of the split colony and the remaining colony
}
\description{
Spit the colony into two new colonies to prevent swarming (in managed populations)
- one colony is with the old queen and a part of the workers and drones (this is the remaining colony)
- the split colony is taken to a new location with part of the workers.
 A new mated queen can be introduced to the split colony.
 If no new queen is introduced, a virgin queen must be present to mate with fathers from DCA and continue colony
}
\examples{
inst/examples/examples_splitColony.R
}
\seealso{
\code{\link[??????]{splitColony}}
}
