% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{supersedeColonies}
\alias{supersedeColonies}
\title{supersedeColonies}
\usage{
\method{supersedeColonies}(colonies, crossVirginQueen, fathers, pWorkers, pDrones)
}
\arguments{
\item{colonies}{AlphaSimRBee Colonies object containing a list of colonies}

\item{pDrones}{}
}
\value{
An updated AlphaSimRBee Colonies object
}
\description{
:  Replicates the process of supersedure, where the
queen is replaced by a new virgin queen. The workers and the drones stay
in the colonies. If no fathers are present, mating of the virgin queen does not occur.
}
\examples{
#Create founder haplotypes
founderPop <- quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP <- SimParam$new(founderPop)

#Create population
base <- newPop(founderPop, simParam=SP)
TODO NOT WORKING EXAMPLE
#Create 10 virgin queen colonies
 apiary1 <- createMultipleMatedColonies(founderPop = base, nColonies = 10, nAvgFathers = 15)

# Select colonies for supersedure

coloniesForSupersedure <- pullColonies(apiary1, p = 0.1)

# Supersedure the colonies

tmp <- supersedeColonies(coloniesForSupersedure)

# Put superseded colonies back to the apiary

apiary1 <- c(apiary1, tmp)

}
