% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{swarmColonies}
\alias{swarmColonies}
\title{swarmColonies}
\usage{
\method{swarmColonies}(colonies, crossVirginQueen, fathers, pWorkers, pDrones)
}
\arguments{
\item{colonies}{AlphaSimRBee Colonies object containing a list of colonies}

\item{pDrones}{}
}
\value{
Two AlphaSim population objects of the swarmed colonies and the remaining colonies
}
\description{
: Replicates the swarming of the colonies - the process in which
a part of the workers leave with the old queen and creates a new colony (the swarm),
while a part of the workers stay with a new queen and the old drones.
The swarming colony contains the old mated queen, a percentage (pSwarm)
of the original colony's workers, no drones and a virgin queen is created from the worker population.
A new location must be given to the new swarm colony.
The colony that stays contains the remaining workers and drones.
A virgin queen is selected from the workers and mated if fathers are present.
}
\examples{
#Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
base = newPop(founderPop, simParam=SP)

#Create 10 virgin queen colonies
 apiary1 = createMultipleMatedColonies(founderPop = base, nColonies = 10, nAvgFathers = 15)

#Build up colonies by adding 1000 workers and 100 drones to each colony in the "colonies" list
 apiary1 = buildUpColonies(apiary1, nWorkers = 1000, nDrones = 100)

 TODO FINISH EXAMPLE

}
