% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L2_Colony.R
\name{swarmColony}
\alias{swarmColony}
\title{Replicates the swarming process and produces two colonies.}
\usage{
\method{swarmColony}(colony, pSwarm, crossVirginQueen. fathers, nWorkers, nDrones, swarmLocation)
}
\arguments{
\item{colony}{Colony class. AlphaSimRBee Colony object from the \code{createColony(...)} call}

\item{pSwarm}{Integer. Percentage of colony that will swarm}

\item{crossVirginQueen}{Logical. Whether a virgin queen is to be mated}

\item{fathers}{AlphaSimR population object. Number of fathers pulled from the DCA}

\item{pWorkers}{Numeric, proportion of workers that are replaced with the workers from the new queen in the remnant colony}

\item{pDrones}{Numeric, proportion of drones that are replaced with the drones from the new queen in the remnant colony}

\item{swarm}{Location Integer. X,Y coordinates of newly made swarmed hive}
}
\value{
Two colonies, one with the new queen and proportion of workers and
one with the old queen and proportion of workers.
}
\description{
List. Replicates the swarming of the colony - the process in which
a part of the workers leave with the old queen and creates a new colony (the swarm),
while a part of the workers stay with a new queen and the old drones.
The swarming colony contains the old mated queen,
 a percentage (pSwarm) of the original colonies workers, no drones and a virgin queen is created from the worker population.
 A new location must be given to the new swarm colony.
 The colony that stays contains the remaining workers and drones. A virgin queen is selected from the workers and mated if fathers are present.
}
\examples{
inst/examples/examples_swarm.R
}
\seealso{
\code{\link[??????]{createColony}}
}
