% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L1_Pop.R
\name{createDCA}
\alias{createDCA}
\title{Creates a drone congregation area (DCA) from the list of colonies}
\usage{
\method{createDCA}(list(colonies))
}
\arguments{
\item{colonies}{A list of colonies, each of AlphaSimRBee Colony object}
}
\value{
Single AlphaSim population object of drones from the provided colonies.
}
\description{
Creates a drone congregation area (DCA) from selected colonies.
The function takes a vector of the colonies and returns a combined population of drones.
}
\examples{
# Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
pop = newPop(founderPop, simParam=SP)

#Creates colony
founderDrones = createFounderDrones(pop[3:200], nDronesPerQueen = 17)
colony1 = createColony(queen = pop[1], fathers = founderDrones[1:17])
colony2 = createColony(queen = pop[2], fathers = founderDrones[18:37])

#Create drones that will be in DCA
colony1@drones = createDrones(colony1, nInd = 300)
colony2@drones = createDrones(colony2, nInd = 300)

DCA = createDCA(c(colony1, colony2))

}
\seealso{
\code{\link[??????]{select_colonies}}
}
