% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L0_auxilary.R
\name{nColonies}
\alias{nColonies}
\title{Number of colonies}
\usage{
nColonies(colonies)
}
\arguments{
\item{colonies}{Colonies}
}
\value{
Integer
}
\description{
Returns the number of colonies in a colonies object
}
\examples{
AlphaSimR
founderGenomes <- quickHaplo(nInd = 3, nChr = 1, segSites = 10)
SP <- SimParam$new(founderGenomes)
basePop <- newPop(founderGenomes)

# Honeybees
drones <- createFounderDrones(pop = basePop[2], nDronesPerQueen = 10)
colony1 <- createColony(queen = basePop[1], fathers = drones[1:5])
colony2 <- createColony(queen = basePop[1], fathers = drones[6:10])
apiary <- c(colony1, colony2)
nColonies(apiary)
nColonies(createColonies(n = 10))

}
