% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{pullColonies}
\alias{pullColonies}
\title{Pull the colonies from the colony list based on IDs.}
\usage{
\method{pullColonies}(colonies,ID, p)
}
\arguments{
\item{colonies}{AlphaSimRBee Colonies object containing a list of colonies}

\item{ID}{IDs of "colony" class objects listed in the "colonies" object}

\item{p}{Percentage of colonies to pull out of the "colonies" list object}
}
\value{
Two lists: a list of selected colonies and an updated inpute colonies
}
\description{
Pull the colonies from the list of all colonies based
on colony IDs and return two lists: a list of selected colonies and
updated original colonies
}
\examples{
#Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
base = newPop(founderPop, simParam=SP)

#Create an apiary with 10 mated colonies (contains only queens, fathers and virgin queens)
apiary1 = createMultipleMatedColonies(base, nColonies = 10, nAvgFathers = 15)

#Build up the colonies to full size by adding workers and drones (drones are 1/10 size of workers)
colonyFullSize = 1000
apiary1 = buildUpColonies(apiary1, nWorkers = colonyFullSize, nDrones = colonyFullSize * 0.1)

#Check colony IDs in colony
apiary@colonies

#Pull colonies with IDs 8, 40 and 45 to create two separate groups of colonies (apiaries)
tmp = pullColonies(apiary1, ID = c(8,40,45))
apiary1 = tmp$remainingColonies
apiary2 = tmp$pulledColonies

}
