% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_L3_Colonies.R
\name{supersedeColonies}
\alias{supersedeColonies}
\title{supersedeColonies}
\usage{
\method{supersedeColonies}(colonies, crossVirginQueen, fathers, pWorkers, pDrones)
}
\arguments{
\item{colonies}{AlphaSimRBee Colonies object containing a list of colonies}

\item{pDrones}{}
}
\value{
An updated AlphaSimRBee Colonies object
}
\description{
:  Replicates the process of supersedure, where the
queen is replaced by a new virgin queen. The workers and the drones stay
in the colonies. If no fathers are present, mating of the virgin queen does not occur.
}
\examples{
#Create founder haplotypes
founderPop = quickHaplo(nInd=200, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)

#Create population
base = newPop(founderPop, simParam=SP)

#Create 10 virgin queen colonies
 apiary1 = createMultipleMatedColonies(founderPop = base, nColonies = 10, nAvgFathers = 15)

#Build up colonies by adding 1000 workers and 100 drones to each colony in the "colonies" list
 apiary1 = buildUpColonies(apiary1, nWorkers = 1000, nDrones = 100)

 TODO FINISH EXAMPLE

}
